Address model artifact signature review comments
